# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "K7 E-Library API",
        "description": "Digital library management system API for educational institutions",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8747/api/v1"
        }
    ],
    "paths": {
...
        "/content": {
            "get": {
                "tags": [
                    "Content"
                ],
                "summary": "List all published content",
                "operationId": "02aac6aeed47c865fa861e805dcd6dff",
                "parameters": [
                    {
                        "name": "filter[type]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ebook",
                                "journal",
                                "student_project",
                                "lecture"
                            ]
                        }
                    },
                    {
                        "name": "filter[language]",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "-created_at",
                                "view_count",
                                "-view_count"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content list"
                    }
                }
            }
        },
        "/content/{slug}": {
            "get": {
                "tags": [
                    "Content"
                ],
                "summary": "Get content by slug",
                "operationId": "a1ed474237392f27c27891d16bbb8cad",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content details"
                    },
                    "404": {
                        "description": "Content not found"
                    }
                }
            }
        },
...
        "/media/upload": {
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Upload file",
                "operationId": "bb13627fbcf5879bebd3ace5a896ed62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "content_item_id"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "content_item_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/media/{id}/download": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Download file",
                "operationId": "4294183454f9df1fef70cd79f061332b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File download"
                    },
                    "404": {
                        "description": "File not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reading-progress/{contentId}": {
            "get": {
                "tags": [
                    "Reading Progress"
                ],
                "summary": "Get reading progress",
                "operationId": "fb5000fef01d19387e96b83a421a8990",
                "parameters": [
                    {
                        "name": "contentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reading progress"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reading Progress"
                ],
                "summary": "Update reading progress",
                "operationId": "69ef09f578261dff3fcd0f5c23e0ed94",
                "parameters": [
                    {
                        "name": "contentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "current_page": {
                                        "type": "integer"
                                    },
                                    "total_pages": {
                                        "type": "integer"
                                    },
                                    "progress_percent": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Progress updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/history": {
            "get": {
                "tags": [
                    "Reading Progress"
                ],
                "summary": "Get reading history",
                "operationId": "d92721665191f60b8ea41165fd813069",
                "responses": {
                    "200": {
                        "description": "Reading history"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/content/{contentId}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get content reviews",
                "operationId": "11a7d780b0bdc9b9d3b0b141242f8cef",
                "parameters": [
                    {
                        "name": "contentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews list"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Submit review",
                "operationId": "9954aee8ec98ca388a7c35790181de99",
                "parameters": [
                    {
                        "name": "contentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "review_text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review submitted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}